Progress: Prec/width for signed ints

+++ ft_printf.h
@@ -58,8 +58,8 @@ void	zero_flags(t_start *start);
 int		escape_check(t_start *start, int *i);
 void	parse_conv_char(t_start *start);
 void	parser(t_start *start);
-int		signed_count(intmax_t n, t_start *start);
-int		unsigned_count(uintmax_t n, int base, t_start *start);
+size_t	signed_count(intmax_t n, t_start *start);
+size_t	unsigned_count(uintmax_t n, int base, t_start *start);
 void	string_pad(t_start *start, int print_len, char *str, char *w_str);
 
 int		print_hex(t_start *start);
